Fix literal type

+++ src/multio/action/statistics-mtg2/cfg/StatisticsConfiguration.cc
@@ -145,7 +145,7 @@ eckit::DateTime StatisticsConfiguration::computeEpoch() const {
 
 
 eckit::DateTime StatisticsConfiguration::computeCurr() const {
-    return epoch() + static_cast<eckit::Second>(std::max(step_, 0L) * timeStep_);
+    return epoch() + static_cast<eckit::Second>(std::max(step_, static_cast<int64_t>(0)) * timeStep_);
 }
